# Cuda memcpy not used