The CUDA memcpy is never called when running the benchmark. 
For the compression, it calls `lzbench_cuda_return_0` which does nothing, causing the output to be set equal [by this part](https://github.com/inikep/lzbench/blob/d138844ea56b36ff1c1c43b259c866069deb64ad/_lzbench/lzbench.cpp#L284C1-L289C10).

During the decompression is the size of the compressed data the same size as the input data, which causes the decompression to use `memcpy` instead of the `lzbench_cuda_memcpy` function. [See this part](https://github.com/inikep/lzbench/blob/d138844ea56b36ff1c1c43b259c866069deb64ad/_lzbench/lzbench.cpp#L311C6-L319).

This also affects the normal memcpy, which is never executed. It is actually executing [this line](https://github.com/inikep/lzbench/blob/d138844ea56b36ff1c1c43b259c866069deb64ad/_lzbench/lzbench.cpp#L313C9-L313C9).

From:
```
lzbench 1.8 (64-bit Linux)  11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
Assembled by P.Skibinski

Compressor name         Compress. Decompress. Compr. size  Ratio Filename
memcpy                  32299 MB/s 32091 MB/s    10192446 100.00 ../datasets/silesia/dickens
cudaMemcpy              34108 MB/s 32113 MB/s    10192446 100.00 ../datasets/silesia/dickens
nvcomp_lz4 1.2.2 -0       627 MB/s  1704 MB/s     6890640  67.61 ../datasets/silesia/dickens
nvcomp_lz4 1.2.2 -1       398 MB/s  1178 MB/s     6688442  65.62 ../datasets/silesia/dickens
```
![before](https://github.com/inikep/lzbench/assets/11787137/a4b56b1c-dc4a-4aa1-84b1-b8abd72ffeae)


To:
```
lzbench 1.8 (64-bit Linux)  11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
Assembled by P.Skibinski

Compressor name         Compress. Decompress. Compr. size  Ratio Filename
memcpy                  17350 MB/s 34845 MB/s    10192446 100.00 ../datasets/silesia/dickens
cudaMemcpy               2810 MB/s  3066 MB/s    10192446 100.00 ../datasets/silesia/dickens
nvcomp_lz4 1.2.2 -0       627 MB/s  1698 MB/s     6890640  67.61 ../datasets/silesia/dickens
nvcomp_lz4 1.2.2 -1       398 MB/s  1176 MB/s     6688442  65.62 ../datasets/silesia/dickens
```
![after](https://github.com/inikep/lzbench/assets/11787137/b93fea88-e131-49b9-a4f0-afca5a3abc0c)

Note that there is something weird going on with memcpy now.