@tansy 2024-01-21
It of curiosity - you called it _'always call the decompressor if the compression was successful'_ but you never check it. You only store this result in a vector.

And also, why it's memcpy compress half of the speed of decompress? I thought it was the same copying data, so it should be any different.